Build fix for linux (Fedora 11 on x86_64).
The malloc_test unit uses dlsym with RTLD_NEXT to find the libc's malloc
function and substitute if no replacement is linked in. However, this
define is a GNU extension, and so the source must define _GNU_SOURCE
before including dlfcn.h.

The dlsym function and friends also require linking with -ldl on Linux.